Consolidate collect them all presenter tests
