--- EurofurenceTests/Application/Components/Collect-them-All/Presenter Tests/BeforeCollectThemAllSceneLoads_PresenterShould.swift
@@ -1,12 +1,0 @@
-import Eurofurence
-import EurofurenceModel
-import XCTest
-
-class BeforeCollectThemAllSceneLoads_PresenterShould: XCTestCase {
-
-    func testNotTellTheSceneToLoadTheGameRequest() {
-        let context = CollectThemAllPresenterTestBuilder().build()
-        XCTAssertNil(context.scene.capturedURLRequest)
-    }
-
-}
