--- EurofurenceTests/Application/Components/Collect-them-All/Presenter Tests/WhenCollectThemAllSceneLoads_PresenterShould.swift
@@ -1,14 +1,0 @@
-import Eurofurence
-import EurofurenceModel
-import XCTest
-
-class WhenCollectThemAllSceneLoads_PresenterShould: XCTestCase {
-
-    func testTellTheSceneToLoadTheURLFromTheCollectThemAllService() {
-        let context = CollectThemAllPresenterTestBuilder().build()
-        context.simulateSceneDidLoad()
-
-        XCTAssertEqual(context.service.urlRequest, context.scene.capturedURLRequest)
-    }
-
-}
